refactor(provider-routing): extract ProviderStatus component, compact tickets-db DDL
- Extract Provider-Status widget into ProviderStatus.svelte (FactoryFloor: 527→490 lines)
- Move interfaces to factory-floor.ts imports (eliminate duplication)
- Extract provider routing DDL into initProviderRouting() helper
- Update quality baseline for tickets-db.ts (1088→1106 lines)

Quality gates S1-S4: no new violations.
