+++ docs/code-quality/baseline.json
@@ -121,8 +121,8 @@
   "S1:website/src/layouts/AdminLayout.astro": {
     "gate": "S1",
     "path": "website/src/layouts/AdminLayout.astro",
-    "metric": 443,
-    "detail": "443 lines > 400 limit (.astro)",
+    "metric": 442,
+    "detail": "442 lines > 400 limit (.astro)",
     "frozen_at": "a938f77f"
   },
   "S1:website/src/lib/coaching-db.ts": {
@@ -149,8 +149,8 @@
   "S1:website/src/lib/tickets-db.ts": {
     "gate": "S1",
     "path": "website/src/lib/tickets-db.ts",
-    "metric": 1088,
-    "detail": "1088 lines > 600 limit (.ts)",
+    "metric": 1106,
+    "detail": "1106 lines > 600 limit (.ts)",
     "frozen_at": "badfd86b"
   },
   "S1:website/src/lib/tickets/admin.ts": {
@@ -550,13 +550,6 @@
     "path": "website/src/pages/api/admin/questionnaires/assignments/[id]/reassign.test.ts",
     "metric": 1,
     "detail": "hardcoded host: web.mentolder.de",
-    "frozen_at": "a938f77f"
-  },
-  "S4:scripts/brainstorm-extract-choice.sh": {
-    "gate": "S4",
-    "path": "scripts/brainstorm-extract-choice.sh",
-    "metric": 1,
-    "detail": "no reference to 'brainstorm-extract-choice.sh' in any configured source",
     "frozen_at": "a938f77f"
   },
   "S4:scripts/build-srgb-icc.mjs": {
